backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
crewai>=0.30.0
langchain>=0.1.10
networkx==3.2
numpy==1.24.3
pandas==2.1.1
scipy==1.11.3
yfinance==0.2.31
kafka-python==2.0.2
neo4j==5.14.0
rdflib==7.0.0
textblob==0.17.1
pydantic==2.4.2
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
httpx==0.25.0
anthropic==0.7.0
graphviz==0.20.1
arch==5.3.1
scikit-learn==1.3.1
matplotlib==3.7.2
seaborn==0.12.2
hmmlearn==0.3.0
pomegranate==0.14.8
